Add files via upload
CRUD is a fundamental concept in computer programming, databases and application design. It stands for Create, Read, Update, Delete, which are the four basic operations that a model should able to do. CRUD Apps are used to interact with databases in a controlled way, and are the basis of data management in many platforms.